Parser finished!

+++ memory.c
@@ -4,7 +4,7 @@
 static void * objs[4096]; // FIXME
 static int top = 0;
 
-void * alloc(unsigned long size) {
+void * alloc(size_t size) {
 	assert(top < 4096);
 	return objs[top++] = malloc(size);
 }
